requirements.txt
asgiref==3.8.1
attrs==25.3.0
boolean.py==4.0
CacheControl==0.14.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cyclonedx-python-lib==8.9.0
defusedxml==0.7.1
Django==5.1.7
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
filelock==3.18.0
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
license-expression==30.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
packageurl-python==0.16.0
packaging==24.2
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.8.0
platformdirs==4.3.7
psycopg2-binary==2.9.10
py-serializable==1.1.2
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
toml==0.10.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
